+++ server/api/archive.go
@@ -16,6 +16,13 @@ const (
 	archiveExtension = ".boardarchive"
 )
 
+func (a *API) registerAchivesRoutes(r *mux.Router) {
+	// Archive APIs
+	r.HandleFunc("/boards/{boardID}/archive/export", a.sessionRequired(a.handleArchiveExportBoard)).Methods("GET")
+	r.HandleFunc("/teams/{teamID}/archive/import", a.sessionRequired(a.handleArchiveImport)).Methods("POST")
+	r.HandleFunc("/teams/{teamID}/archive/export", a.sessionRequired(a.handleArchiveExportTeam)).Methods("GET")
+}
+
 func (a *API) handleArchiveExportBoard(w http.ResponseWriter, r *http.Request) {
 	// swagger:operation GET /boards/{boardID}/archive/export archiveExportBoard
 	//
@@ -70,75 +77,6 @@ func (a *API) handleArchiveExportBoard(w http.ResponseWriter, r *http.Request) {
 	opts := model.ExportArchiveOptions{
 		TeamID:   board.TeamID,
 		BoardIDs: []string{board.ID},
-	}
-
-	filename := fmt.Sprintf("archive-%s%s", time.Now().Format("2006-01-02"), archiveExtension)
-	w.Header().Set("Content-Type", "application/octet-stream")
-	w.Header().Set("Content-Disposition", "attachment; filename="+filename)
-	w.Header().Set("Content-Transfer-Encoding", "binary")
-
-	if err := a.app.ExportArchive(w, opts); err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-	}
-
-	auditRec.Success()
-}
-
-func (a *API) handleArchiveExportTeam(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation GET /teams/{teamID}/archive/export archiveExportTeam
-	//
-	// Exports an archive of all blocks for all the boards in a team.
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: teamID
-	//   in: path
-	//   description: Id of team
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     content:
-	//       application-octet-stream:
-	//         type: string
-	//         format: binary
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-	if a.MattermostAuth {
-		a.errorResponse(w, r.URL.Path, http.StatusNotImplemented, "not permitted in plugin mode", nil)
-	}
-
-	vars := mux.Vars(r)
-	teamID := vars["teamID"]
-
-	ctx := r.Context()
-	session, _ := ctx.Value(sessionContextKey).(*model.Session)
-	userID := session.UserID
-
-	auditRec := a.makeAuditRecord(r, "archiveExportTeam", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelRead, auditRec)
-	auditRec.AddMeta("TeamID", teamID)
-
-	boards, err := a.app.GetBoardsForUserAndTeam(userID, teamID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-	ids := []string{}
-	for _, board := range boards {
-		ids = append(ids, board.ID)
-	}
-
-	opts := model.ExportArchiveOptions{
-		TeamID:   teamID,
-		BoardIDs: ids,
 	}
 
 	filename := fmt.Sprintf("archive-%s%s", time.Now().Format("2006-01-02"), archiveExtension)
@@ -223,5 +161,74 @@ func (a *API) handleArchiveImport(w http.ResponseWriter, r *http.Request) {
 	}
 
 	jsonStringResponse(w, http.StatusOK, "{}")
+	auditRec.Success()
+}
+
+func (a *API) handleArchiveExportTeam(w http.ResponseWriter, r *http.Request) {
+	// swagger:operation GET /teams/{teamID}/archive/export archiveExportTeam
+	//
+	// Exports an archive of all blocks for all the boards in a team.
+	//
+	// ---
+	// produces:
+	// - application/json
+	// parameters:
+	// - name: teamID
+	//   in: path
+	//   description: Id of team
+	//   required: true
+	//   type: string
+	// security:
+	// - BearerAuth: []
+	// responses:
+	//   '200':
+	//     description: success
+	//     content:
+	//       application-octet-stream:
+	//         type: string
+	//         format: binary
+	//   default:
+	//     description: internal error
+	//     schema:
+	//       "$ref": "#/definitions/ErrorResponse"
+	if a.MattermostAuth {
+		a.errorResponse(w, r.URL.Path, http.StatusNotImplemented, "not permitted in plugin mode", nil)
+	}
+
+	vars := mux.Vars(r)
+	teamID := vars["teamID"]
+
+	ctx := r.Context()
+	session, _ := ctx.Value(sessionContextKey).(*model.Session)
+	userID := session.UserID
+
+	auditRec := a.makeAuditRecord(r, "archiveExportTeam", audit.Fail)
+	defer a.audit.LogRecord(audit.LevelRead, auditRec)
+	auditRec.AddMeta("TeamID", teamID)
+
+	boards, err := a.app.GetBoardsForUserAndTeam(userID, teamID)
+	if err != nil {
+		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
+		return
+	}
+	ids := []string{}
+	for _, board := range boards {
+		ids = append(ids, board.ID)
+	}
+
+	opts := model.ExportArchiveOptions{
+		TeamID:   teamID,
+		BoardIDs: ids,
+	}
+
+	filename := fmt.Sprintf("archive-%s%s", time.Now().Format("2006-01-02"), archiveExtension)
+	w.Header().Set("Content-Type", "application/octet-stream")
+	w.Header().Set("Content-Disposition", "attachment; filename="+filename)
+	w.Header().Set("Content-Transfer-Encoding", "binary")
+
+	if err := a.app.ExportArchive(w, opts); err != nil {
+		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
+	}
+
 	auditRec.Success()
 }
